+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuProductsTest.java
@@ -1,9 +1,7 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import kitchenpos.menus.application.FakeProductClient;
 import kitchenpos.menus.dto.MenuProductCreateRequest;
-import kitchenpos.menus.tobe.domain.menu.MenuProducts;
-import kitchenpos.menus.tobe.domain.menu.ProductClient;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
